+++ src/test/ExSln3/c99/gen/board_IGeneralBoard.c
@@ -11,7 +11,7 @@
 /// Button that player presses to trap the light beam.
 /// </summary>
 /// <returns></returns>
-IDigIn board_IGeneralBoard_get_trap_button(board_IGeneralBoard * self)
+hal_IDigIn * board_IGeneralBoard_get_trap_button(board_IGeneralBoard * self)
 {
     return self->obj_vtable->get_trap_button(self->obj);
 }
@@ -21,13 +21,13 @@ IDigIn board_IGeneralBoard_get_trap_button(board_IGeneralBoard * self)
 /// Button that player presses to start the game.
 /// </summary>
 /// <returns></returns>
-IDigIn board_IGeneralBoard_get_start_button(board_IGeneralBoard * self)
+hal_IDigIn * board_IGeneralBoard_get_start_button(board_IGeneralBoard * self)
 {
     return self->obj_vtable->get_start_button(self->obj);
 }
 
 
-IDigOut board_IGeneralBoard_get_main_led(board_IGeneralBoard * self)
+hal_IDigOut * board_IGeneralBoard_get_main_led(board_IGeneralBoard * self)
 {
     return self->obj_vtable->get_main_led(self->obj);
 }
@@ -36,7 +36,7 @@ IDigOut board_IGeneralBoard_get_main_led(board_IGeneralBoard * self)
 //DigInArray get_settings_switches();
 //DigOutArray get_sweep_leds();
 
-u32 board_IGeneralBoard_get_time_ms(board_IGeneralBoard * self)
+uint32_t board_IGeneralBoard_get_time_ms(board_IGeneralBoard * self)
 {
     return self->obj_vtable->get_time_ms(self->obj);
 }